Drops unnecessary type casting

+++ src/unserialize.ts
@@ -49,7 +49,7 @@ function unserializeItem(parser: Parser, scope: Record<string, any>, options: Op
     const value = parser.readUntil(';')
     let parsedValue: number | BigInt = type === 'int' ? parseInt(value, 10) : parseFloat(value)
     if (parsedValue.toString() !== value) {
-      parsedValue = BigInt(value) as BigInt
+      parsedValue = BigInt(value)
     }
     return parsedValue
   }
